refactor: Separate interfaces into different entity files for better organization
