+++ src/openai/dto/entities/chat.interface.ts
@@ -1,0 +1,7 @@
+import { Message } from './message.interface';
+
+export interface Chat {
+    chatId: number;
+    customerId: number;
+    messages: Message[];
+}
