+++ src/openai/dto/entities/message.interface.ts
@@ -1,0 +1,5 @@
+export interface Message {
+    name: string;
+    content: string;
+    role: Role;
+}
