+++ src/openai/dto/entities/role.enum.ts
@@ -1,0 +1,6 @@
+enum Role {
+    ASSISTANT = 'assistant',
+    USER = 'user',
+    SYSTEM = 'system',
+    FUNCTION = 'function',
+}
