+++ src/openai/dto/entities/user.interface.ts
@@ -1,0 +1,5 @@
+export interface User {
+    userId: number;
+    username: string;
+    roles?: Role[];
+}
